4. Using "const" Widgets & Constructors to Improve the App Performance:
You can skip unnecessary widget rebuilds if you mark a widget as const.
This object will never ever change, ant that tells Flutter that since
this will never change, when it rebuilds the widget tree, for this
specific widget it can simply take the old widget which was in the same
position, it doesn't need to recreate the object.